Rename `markdown.marp.toggleMarpPreview` to `markdown.marp.toggleMarpFeature`
